requirements.txt
pandas==1.5.0
selenium==4.5.0
python-dotenv==0.21.0